Create Media componant used by detail and pois
