+++ src/app/media/directives.js
@@ -1,0 +1,15 @@
+'use strict';
+
+function mediaPlayer() {
+    return {
+        restrict: 'E',
+        replace: true,
+        scope: true,
+        template: require('./templates/media-modal.html'),
+        controller: 'MediaController'
+    };
+}
+
+module.exports = {
+    mediaPlayer: mediaPlayer
+};
