+++ src/app/media/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+angular.module('rando.media', [])
+    .controller('MediaController', require('./controllers').MediaController)
+    .directive('mediaPlayer', require('./directives').mediaPlayer);
