+++ src/app/pois/controllers.js
@@ -96,7 +96,7 @@ function PoisListeController($scope, $rootScope, globalSettings, utilsFactory, $
 
     $scope.openPlayer = function openPlayer (media) {
         $modal.open({
-            template: require('./templates/media-modal.html'),
+            template: require('../media/templates/media-modal.html'),
             controller: 'MediaController',
             resolve: {
                 media: function () {
@@ -114,28 +114,7 @@ function PoisListeController($scope, $rootScope, globalSettings, utilsFactory, $
 
 }
 
-function MediaController(media, $scope, $timeout, $modalInstance) {
-    function initPlayer() {
-        var player = document.createElement('iframe');
-        player.width = '100%';
-        player.height = '100%';
-        player.frameborder = '0';
-        player.setAttribute('allowfullscreen', '');
-        player.setAttribute('webkitallowfullscreen', '');
-        player.setAttribute('mozallowfullscreen', '');
-        player.src = media.url;
-        document.querySelector('#media-player').appendChild(player);
-
-    }
-
-    $scope.close = function close () {
-        $modalInstance.dismiss('close');
-    };
-
-    $timeout(initPlayer, 500);
-}
 
 module.exports = {
-    PoisListeController: PoisListeController,
-    MediaController: MediaController
+    PoisListeController: PoisListeController
 };
